pyproject.toml
[project]
name = "qq-group-bot"
version = "0.1.0"
description = "A starter NoneBot2 QQ group bot powered by LLOneBot."
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "nonebot2[fastapi]<3.0.0,>=2.4.3",
    "nonebot-adapter-onebot<3.0.0,>=2.4.6",
    "nonebot-plugin-apscheduler<1.0.0,>=0.5.0",
    "nonebot-plugin-analysis-bilibili>=2.8.1",
    "nonebot-plugin-send-anything-anywhere>=0.7.1",
    "nonebot_plugin_withdraw>=0.4.1",
    "nonebot-plugin-whateat-pic>=1.5.7",
    "nonebot-plugin-today-waifu>=0.1.10",
    "httpx<1.0,>=0.27",
    "nonebot-plugin-wordsnorote>=0.0.2",
    "nonebot-plugin-reboot>=0.1.4",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-analysis-bilibili = ["nonebot_plugin_analysis_bilibili"]
nonebot-plugin-send-anything-anywhere = ["nonebot_plugin_saa"]
nonebot_plugin_withdraw = ["nonebot_plugin_withdraw"]
nonebot-plugin-whateat-pic = ["nonebot_plugin_whateat_pic"]
nonebot-plugin-today-waifu = ["nonebot_plugin_today_waifu"]
nonebot-plugin-wordsnorote = ["nonebot_plugin_wordsnorote"]
nonebot-plugin-reboot = ["nonebot_plugin_reboot"]

[tool.nonebot.adapters]
"@local" = []
nonebot-adapter-onebot = [{name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11"}]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
